+++ packages/components/src/DataTable/DataTable.css.d.ts
@@ -7,10 +7,6 @@ declare const styles: {
   readonly "pinFirstHeaderSortable": string;
   readonly "sortableColumn": string;
   readonly "clickableRow": string;
-  readonly "sortUp": string;
-  readonly "topArrow": string;
-  readonly "sortDown": string;
-  readonly "bottomArrow": string;
 };
 export = styles;
