+++ packages/components/src/DataTable/SortIcon.css.d.ts
@@ -1,0 +1,7 @@
+declare const styles: {
+  readonly "sortIcon": string;
+  readonly "topArrow": string;
+  readonly "bottomArrow": string;
+};
+export = styles;
+
