--- packages/components/src/DataTable/constants/SortIcon.tsx
@@ -1,24 +1,0 @@
-import React from "react";
-
-export function SortIcon(): JSX.Element {
-  return (
-    <svg
-      width="10"
-      height="14"
-      viewBox="0 0 10 14"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        id="topArrow"
-        d="M8.99999 5.66609C8.60946 6.05662 7.9763 6.05662 7.58577 5.66609L4.99615 3.14264L2.4142 5.66609C2.02367 6.05662 1.39051 6.05662 0.999986 5.66609C0.609461 5.27557 0.609461 4.6424 0.999986 4.25188L4.29288 0.958983C4.6834 0.568459 5.31657 0.56846 5.70709 0.958984L8.99999 4.25188C9.39051 4.6424 9.39051 5.27557 8.99999 5.66609Z"
-        fill="#657884"
-      />
-      <path
-        id="bottomArrow"
-        d="M1.00001 8.33391C1.39054 7.94338 2.0237 7.94338 2.41423 8.33391L5.00384 10.8574L7.5858 8.33391C7.97633 7.94338 8.60949 7.94338 9.00001 8.33391C9.39054 8.72443 9.39054 9.3576 9.00001 9.74812L5.70712 13.041C5.3166 13.4315 4.68343 13.4315 4.29291 13.041L1.00001 9.74812C0.609489 9.3576 0.60949 8.72443 1.00001 8.33391Z"
-        fill="#657884"
-      />
-    </svg>
-  );
-}
